Adding a configuration option to inhibit suspend and idle behavior.
Two boolean flags are added to the dconf:
  * inhibit-suspend
  * inhibit-idle

A graphical combobox is added to the settings dialog where 4 options are possible:
  * none
  * idle
  * suspend
  * both

Whenever something other than none is selected the powermanagement is adjusted during playback and resetted afterwards.

The already existing code for the fullscreen inhibit is adjusted to also use the inhibitor.

Implements: https://github.com/gnumdk/lollypop/issues/1188